backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
youtube-transcript-api>=0.6.3
pydantic>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0